Update prompt format and move imageStorage.py
Prompt was previously making LLM return the detected topics and used files in a '$' separated format near the end. The new prompt creates a separate identifying string for both topics and files detected.

imageStorage.py was not able to be imported within views.py due to it being within mySite folder. It is now in the mySite folder.